PR comments and lints
