+++ crates/sui-types/src/messages_consensus.rs
@@ -1,8 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::authenticator_state::ActiveJwk;
-use crate::base_types::{AuthorityName, ObjectRef, SequenceNumber, TransactionDigest};
+use crate::base_types::{AuthorityName, ObjectRef, TransactionDigest};
 use crate::messages_checkpoint::{
     CheckpointSequenceNumber, CheckpointSignatureMessage, CheckpointTimestamp,
 };
@@ -26,52 +25,6 @@ pub struct ConsensusCommitPrologue {
     pub round: u64,
     /// Unix timestamp from consensus
     pub commit_timestamp_ms: CheckpointTimestamp,
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Clone, Serialize, Deserialize)]
-pub struct AuthenticatorStateCreate {
-    // note: epoch is not used, it exists just to give the transaction a guaranteed
-    // unique digest.
-    /// Epoch of the authenticator state create transaction
-    pub epoch: u64,
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Clone, Serialize, Deserialize)]
-pub struct AuthenticatorStateExpire {
-    // note: epoch is not used, it exists just to give the transaction a guaranteed
-    // unique digest.
-    /// Epoch of the authenticator state expiry transaction
-    pub epoch: u64,
-    /// expire JWKs that have a lower epoch than this
-    pub min_epoch: u64,
-    /// The initial version of the authenticator object that it was shared at.
-    pub authenticator_obj_initial_shared_version: SequenceNumber,
-}
-
-impl AuthenticatorStateExpire {
-    pub fn authenticator_obj_initial_shared_version(&self) -> SequenceNumber {
-        self.authenticator_obj_initial_shared_version
-    }
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Clone, Serialize, Deserialize)]
-pub struct AuthenticatorStateUpdate {
-    /// Epoch of the authenticator state update transaction
-    pub epoch: u64,
-    /// Consensus round of the authenticator state update
-    pub round: u64,
-    /// newly active jwks
-    pub new_active_jwks: Vec<ActiveJwk>,
-    /// The initial version of the authenticator object that it was shared at.
-    pub authenticator_obj_initial_shared_version: SequenceNumber,
-    // to version this struct, do not add new fields. Instead, add a AuthenticatorStateUpdateV2 to
-    // TransactionKind.
-}
-
-impl AuthenticatorStateUpdate {
-    pub fn authenticator_obj_initial_shared_version(&self) -> SequenceNumber {
-        self.authenticator_obj_initial_shared_version
-    }
 }
 
 // In practice, JWKs are about 500 bytes of json each, plus a bit more for the ID.
